requirements.txt
Django
psycopg2-binary
stripe
django-environ
djangorestframework
asgiref
python-dotenv
